requirements.txt
fastapi
uvicorn
openai
gspread
google-auth
python-dotenv
